requirements.txt
pandas
pymongo
streamlit
python-dotenv
plotly
dnspython
chardet
